Add a way to set the Kind from Embeddable #164
We have this use case where we use a Wrapper that embed a generic
object in one of its field.

```
@Entity
public class Wrapper {

   @Embedded
   private Interface content;

   ...
}
```

The content can be of any subtype of Interface (some of them will be
outside of our control) and the ```Kind``` value can change with any of
these subtypes.

This Pull Request allow a developer to provide an ```Embeddable``` class
that will override the ```Kind``` of the enclosing ```Entity```.

It will look like this:

```
@Embeddable
@OverrideKind(kind = "mycontent")
public class MyContent implement Interface {

   ...
}
```